connection: ssh: Use individuals' host variables for physical host IP
It appears that since Ansible 2.4 the 'hostvars' variable does not
contain the host variables set by the linear plugin. However, the
physical_host_addrs variable set by that plugin is already present on
the host we are interested in. As such we can obtain that variable by
simply using the get_vars() method on the Host object.

Change-Id: Ibe2cff5fabf23e5ba6d5d165fb5406bca690a1e2