Fix delegation to hosts not in inventory
When using Ansible 2.5, delegation of a task from a container to a host
that is not mentioned in the inventory is currently broken.

The connection plugin works by comparing the host vars 'container_name'
and 'physical_host' to determine if a target host is a container. Both
of these variables are specific to hosts within an OSA generated
inventory. When delegating to a host outside of the inventory however,
the variables will remain cached from the originating host, so a task
wouldn't actually delegate at all.

This is fixed by checking that the 'inventory_hostname' is the same as
the 'container_name'.

Change-Id: I69f2eed35859bdc149e5ed21441eab7c8a8352cf