Build skel to separate the OSA connection plugin
This change implements the basic skeleton to create a new repo for the
OSA connection plugin. This change will allow the connection plugin to
be installed standalone, and maintained by a larger community.

This change has pruned the git history so that all of the old history
and commiters record for the connection plugin remains intact.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>